requirements.txt
python-dotenv
requests
msal
pandas
openpyxl
# Optional, for Windows corporate cert store trust
certifi-win32
openai
